move event listeners in api crate
