+++ api/Cargo.toml
@@ -105,3 +105,4 @@ assert_matches = "1.5"
 testcontainers = "0.14.0"
 tracing-subscriber = "0.3"
 jsonwebtoken = "8.3.0"
+mockito = "0.31.0"
