+++ api/Procfile
@@ -1,4 +1,5 @@
 web: ROCKET_PORT=$PORT ./target/release/api
+event-listeners: ./target/release/listeners
 events_sanity_checks: RUST_LOG=info ./target/release/events_sanity_checks
 quotes_syncer: RUST_LOG=info ./target/release/quotes_syncer
 hasura: ./hasura/hasura.sh
