+++ api/src/domain/listeners/quote_syncer.rs
