+++ event-listeners/Cargo.toml
@@ -77,7 +77,6 @@ presentation = { path = "../common/presentation" }
 async-std = { version = "1.12.0", features = ["attributes"] }
 assert_matches = "1.5"
 envtestkit = "1.1.2"
-mockito = "0.31.0"
 rstest = "0.15.0"
 testing = { path = "../common/testing" }
 rust_decimal_macros = "1.26"
