+++ event-listeners/tests/fixtures/issues.rs
@@ -6,7 +6,7 @@ use event_listeners::models;
 use infrastructure::database::schema::github_issues;
 
 use super::*;
-use crate::context::github_indexer::Context;
+use crate::context::Context;
 
 pub fn x1061() -> GithubIssue {
 	GithubIssue {
