+++ event-listeners/tests/fixtures/reviews.rs
@@ -8,7 +8,7 @@ use event_listeners::models;
 use infrastructure::database::schema::github_pull_request_reviews;
 
 use super::*;
-use crate::context::github_indexer::Context;
+use crate::context::Context;
 
 pub fn approved(
 	pull_request_id: GithubPullRequestId,
