+++ event-listeners/tests/github_indexing_it.rs
@@ -14,7 +14,7 @@ use olog::info;
 use rstest::rstest;
 use testcontainers::clients::Cli;
 
-use crate::context::{docker, github_indexer::Context};
+use crate::context::{docker, Context};
 
 mod context;
 mod fixtures;
