+++ event-listeners/tests/on_demand_indexing_it.rs
@@ -6,7 +6,7 @@ use rstest::rstest;
 use testcontainers::clients::Cli;
 
 use crate::{
-	context::{docker, github_indexer::Context},
+	context::{docker, Context},
 	fixtures::*,
 };
