--- event-listeners/tests/resources/wiremock/coinmarketcap/mappings/cryptocurrency-quotes.json
@@ -1,15 +1,0 @@
-{
-    "request": {
-        "url": "/v2/cryptocurrency/quotes/latest?id=1027&convert=USD",
-        "method": "GET",
-        "headers": {
-            "X-CMC_PRO_API_KEY": {
-                "contains": "CMC-API-KEY"
-            }
-        }
-    },
-    "response": {
-        "status": 200,
-        "bodyFileName": "body-cryptocurrency-quotes.json"
-    }
-}
